requirements.txt
Flask==3.1.0
python-dotenv==1.0.1
pandas==2.2.3
numpy==1.26.4
scipy==1.14.1
scikit-learn==1.5.2
xgboost==2.1.3
praw==7.8.1
vaderSentiment==3.3.2
requests==2.32.3
gunicorn==23.0.0
